fixed possible controller problem when converting between Integer and int

+++ Engine/src/input/Controller.java
@@ -62,7 +62,7 @@ public abstract class Controller {
 			byte value = buttons.get();
 			Integer action = controllerBinding.get(button);
 			if (action != null) {
-				Keyboard.setKey(action, value);
+				Keyboard.setKey(action.intValue(), value);
 			}
 			button++;
 		}
